+++ scoda/utils/enumerations.py
@@ -4,8 +4,10 @@ import enum
 
 
 class TokeniserType(enum.Enum):
-    NOTELIKE_TOKENISER = "notelike_tokeniser"
     MIDILIKE_TOKENISER = "midilike_tokeniser"
+    STANDARD_NOTELIKE_TOKENISER = "standard_notelike_tokeniser"
+    COF_NOTELIKE_TOKENISER = "cof_notelike_tokeniser"
+    LARGE_DICTIONARY_NOTELIKE_TOKENISER = "large_dictionary_notelike_tokeniser"
     GRIDLIKE_TOKENISER = "gridlike_tokeniser"
     TRANSPOSED_NOTELIKE_TOKENISER = "transposed_notelike_tokeniser"